[2/2] CMParts: On-The-Go mode
On-The-Go turns on the camera of your phone and streams it to an overlay,
which is drawn over everything.
this makes you feel like your display is transparent, as you see the content
behind it.

its really useful when you are using your phone while walking, thats why
we called it "On-The-Go"

    have fun!

Change-Id: Iec8a2494f7dfcece2890eb0c6a1f93ad452644fc
Signed-off-by: scotthartbti <scott.hart.bti@gmail.com>